refactor: types and schemas.
