+++ src/components/pages/Vota.astro
@@ -1,10 +1,13 @@
 ---
 import Layout from "@/layouts/Layout.astro"
 import { VoteSystem } from "@/components/VoteSystem.tsx"
+
 import Button from "@/components/Button.astro"
 import { getSession } from "auth-astro/server"
 import Svg from "@/components/Svg.astro"
 import Img from "@/components/Img.astro"
+
+const { i18n } = Astro.locals;
 
 const session = await getSession(Astro.request)
 
@@ -25,75 +28,55 @@ const bkg = session ? "vote-system-bkg.webp" : "vota-bkg.webp"
       !session && (
         <div class="mx-auto flex flex-col max-w-7xl px-6 pt-40">
           <h1 class="uppercase mb-10 text-left text-3xl lg:text-5xl font-bold tracking-wider leading-loose max-w-xl text-balance">
-            Las votaciones ya están aquí
+            {i18n.VOTE.TITLE}
           </h1>
 
           <div class="mb-10">
             <Button id="vote-button">
-              ¡Empezar con las votaciones!
+              {i18n.VOTE.CALL_TO_VOTE}
             </Button>
           </div>
-          <p class="max-w-3xl text-pretty text-2xl px-2 mb-10">
-            Ha llegado el momento más importante.{" "}
-            <strong>Votar en la 3ª Edición de los Premios ESLAND</strong>. Para
-            que tus votos se contabilicen de forma correcta, te dejamos algunos
-            consejos:
-          </p>
+          <p
+            set:html={i18n.VOTE.INTRO_CONTENT}
+            class="max-w-3xl text-pretty text-2xl px-2 mb-10"
+          />
 
           <ul class="flex flex-col gap-y-10 text-2xl pl-4 font-extralight">
             <li class="flex gap-x-6 items-center">
               <Svg src="sort.vote" />
-              <span>
-                Vota, <strong>por orden de preferencia</strong>, hasta 4
-                nominados por categoría.
-              </span>
+              <span set:html={i18n.VOTE.ADVICE_1} />
             </li>
 
             <li class="flex gap-x-6 items-center">
               <Svg src="check.pencil" />
-              <span>
-                <strong>No es obligatorio votar en cada categoría.</strong>
-              </span>
+              <span set:html={i18n.VOTE.ADVICE_2} />
             </li>
 
             <li class="flex gap-x-6 items-center">
               <Svg src="podium" />
-              <span>
-                No es obligatorio pero sí recomendable el
-                <strong>uso del voto rankeado</strong>.
-              </span>
+              <span set:html={i18n.VOTE.ADVICE_3} />
             </li>
 
             <li class="flex gap-x-6 items-center">
               <>
-                <Svg src="tap.button" />
-                <span>
-                  Una vez tengas clara tu elección,{" "}
-                  <strong>asegúrate de pulsar “enviar mis votos”.</strong>
-                </span>
+              <Svg src="tap.button" />
+                <span set:html={i18n.VOTE.ADVICE_4} />
               </>
             </li>
 
             <li class="flex gap-x-6 items-center">
               <>
-                <Svg src="share" />
-                <span>
-                  <strong>Recuerda compartir tus votos</strong> en redes
-                  sociales.
-                </span>
+              <Svg src="share" />
+                <span set:html={i18n.VOTE.ADVICE_5} />
               </>
             </li>
 
             <li class="flex gap-x-6 items-center">
               <Svg src="think.brain" />
-              <span>
-                Elige sabiamente y <strong>revisa tus votos</strong>, solo
-                tienes una oportunidad.
-              </span>
+              <span set:html={i18n.VOTE.ADVICE_6} />
             </li>
             <span class="opacity-85 text-base mt-4">
-              Las votaciones estarán abiertas hasta el domingo 14 de enero a las
-              24:00h (hora española)
+              {i18n.VOTE.TIME_LIMIT}
             </span>
           </ul>
         </div>
@@ -103,7 +86,7 @@ const bkg = session ? "vote-system-bkg.webp" : "vota-bkg.webp"
     {
       session && (
         <div class="mx-auto flex flex-col max-w-7xl pt-20 relative">
-          <VoteSystem client:load>
+          <VoteSystem client:load i18n={i18n}>
             <div class="flex items-center gap-4">
               <img
                 class="w-12 h-12 rounded-full"
@@ -116,7 +99,7 @@ const bkg = session ? "vote-system-bkg.webp" : "vota-bkg.webp"
                   id="logout-button"
                   class="text-sm hover:border-white border-b font-light border-transparent text-white flex justify-center items-center gap-x-2"
                 >
-                  <span>Cerrar sesión</span>
+                  <span>{i18n.LOG_OUT}</span>
                 </button>
               </div>
             </div>
