--- src/types/votes.ts
@@ -1,1 +1,0 @@
-export type Votes = Array<Array<string>>
